+++ rr/ns.js
@@ -9,6 +9,14 @@ class NS extends RR {
     if (!this.validHostname('NS', 'address', opts.address)) return
     this.set('address', opts.address)
   }
+
+  toBind () {
+    return `${this.get('name')} ${this.get('ttl')} ${this.get('class')}  NS  ${this.get('address')}\n`
+  }
+
+  toTinydns () {
+    return `&${this.get('name')}::${this.get('address')}:${this.get('ttl')}:${this.get('timestamp')}:${this.get('location')}\n`
+  }
 }
 
 module.exports = NS